New code for MVC patern
A new version of the program, which takes into account the possibility of incorrect input of a four-digit number with non-repeating numbers. To simplify the further modification of the program, the structure of the program is changed to the MVC pattern.
The Controller class handles user input. At the moment, it still has the logic of writing to the database, but then it will be moved to the Model class.
The Model class produces the basic data processing, namely:
- generates a random four-digit number with non-repeating numbers (in the RandomGenerator class);
- compares two four-digit numbers and calculates the number of matches by digits and place in digits in the number;
- stores the values ​​of bulls and cows in ArrayList <Integer>.
The View class generates messages on the screen. Hereinafter, there will be a GUI or web page code.